% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSPRdata.R
\name{getSPRdata}
\alias{getSPRdata}
\title{Retrieve external systemPipeRdata resources}
\usage{
getSPRdata(bam = FALSE)
}
\arguments{
\item{bam}{Logical. If \code{TRUE}, also retrieve the precomputed BAM
files. Default is \code{FALSE}.}
}
\value{
A list with paths to the FASTQ and annotation directories.
  If \code{bam = TRUE}, the list also contains the BAM directory.
}
\description{
Retrieves and caches the external example resources used by
\pkg{systemPipeRdata}. FASTQ and annotation resources are always made
available. Precomputed BAM files are retrieved only when
\code{bam = TRUE}.
}
\details{
Downloaded archives are managed by \pkg{BiocFileCache}. Extracted
resources are stored in a versioned systemPipeRdata cache directory.
}
\examples{
\dontrun{
paths <- getSPRdata()
paths$fastqdir
paths$annotationdir

paths <- getSPRdata(bam = TRUE)
paths$bamdir
}

}
