\name{pathList}
\alias{pathList}
\title{
Return location of sample data 
}
\description{
Function to return paths to sample data provided by \code{sytemPipeRdata} package.
}
\usage{
pathList()
}
\value{
\code{list}
}
\author{
Thomas Girke
}
\examples{
samplepaths <- pathList()
}
\keyword{ utilities }
